feat. center carousel complete
