+++ src/Carousel/SimpleCarousel/simpleCarousel.module.scss
@@ -1,0 +1,69 @@
+.simpleCarousel {
+  .carousel {
+    width: 940px;
+    position: relative;
+
+    .out {
+      overflow: hidden;
+      width: 800px;
+      margin: 0 70px;
+
+      ul {
+        list-style: none;
+        white-space: nowrap;
+        padding-inline-start: 0;
+        transition: 0.3s;
+
+        .in {
+          display: inline-block;
+          width: 198px;
+          height: 100px;
+          border: 1px solid black;
+          background-color: green;
+          text-align: center;
+          font-size: 30px;
+        }
+      }
+    }
+
+    .moveBtn {
+      position: absolute;
+      border: 1px solid black;
+      border-radius: 25px;
+      background-color: white;
+      cursor: pointer;
+      top: 30%;
+
+      .arrow {
+        width: 10px;
+        height: 10px;
+      }
+
+      .iconLeft {
+        border-top: 2px solid black;
+        border-left: 2px solid black;
+        transform: rotate(-45deg);
+      }
+
+      .iconRight {
+        border-top: 2px solid black;
+        border-right: 2px solid black;
+        transform: rotate(45deg);
+      }
+    }
+
+    .left {
+      left: 10px;
+      padding: 15px 12px 15px 18px;
+    }
+
+    .right {
+      right: 10px;
+      padding: 15px 18px 15px 12px;
+    }
+
+    .hide {
+      display: none;
+    }
+  }
+}
